chore: extract class to new file
- Moved Criteria and Option classes to separate files.
- Updated imports in DraxoGrid for the newly created classes.
